Switch to prefixing component styles
